requirements.txt
ultralytics==8.3.0
streamlit==1.38.0
Pillow==10.4.0
torch>=2.0.0
